change output locations
New base location is a env var $OUTPUT_FOLDER, defaults to `.`.
Folder layout has been changed: instead of directory-per-item, it is now directory-per-playlist. Directory name is the playlist name and id, separated by a `.`.
Filename is changed, and is prefixed with the two-digit track index.

Due to these changes, all files for a playlist are now in the same folder; you can easily listen in order by simply playing files from same directory in order.